fix(): google strategy타입 설정
