--- src/modules/auth/google.passport.strategy.ts
@@ -1,37 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { PassportStrategy } from '@nestjs/passport';
-
-import { Strategy, type VerifyCallback } from 'passport-google-oauth20';
-
-@Injectable()
-export class GoogleStrategy extends PassportStrategy(Strategy, 'google') {
-  constructor() {
-    super({
-      clientID: process.env.OAUTH_GOOGLE_ID || '',
-      clientSecret: process.env.OAUTH_GOOGLE_SECRET || '',
-      callbackURL: process.env.OAUTH_GOOGLE_REDIRECT || '',
-      scope: ['email', 'profile'],
-      passReqToCallback: false,
-    });
-  }
-
-  async validate(
-    accessToken: string,
-    refreshToken: string,
-    profile: any,
-    done: VerifyCallback,
-  ): Promise<any> {
-    const { name, emails, photos } = profile;
-    const user = {
-      email: emails[0]?.value,
-      firstName: name.givenName,
-      lastName: name.familyName,
-      picture: photos[0]?.value,
-      accessToken,
-      refreshToken,
-    };
-    done(null, user);
-
-    return user;
-  }
-}
